MAINT: Replace deprecated uses of numpy.{float,int} (#512)
* Replace uses of np.{float,int} with {float,int}

* Replace use of np.float in .stats.stat_density.compute_density()

+++ plotnine/guides/guide_legend.py
@@ -305,8 +305,8 @@ class guide_legend(guide):
         # labels
         labels = []
         for item in self.key['label']:
-            if isinstance(item, np.float) and np.float.is_integer(item):
-                item = np.int(item)  # 1.0 to 1
+            if isinstance(item, float) and float.is_integer(item):
+                item = int(item)  # 1.0 to 1
             va = 'center' if self.label_position == 'top' else 'baseline'
             ta = TextArea(item, textprops=dict(color='black', va=va))
             labels.append(ta)
